requirements.txt
streamlit
openai
python-dotenv
pandas
pdfplumber
faiss-cpu
requests
fuzzywuzzy
python-Levenshtein
